requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.4
bleach==3.3.0
CacheControl==0.12.6
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
charset-normalizer==2.0.3
colorama==0.4.4
contextlib2==0.6.0.post1
cryptography==3.3.2
defusedxml==0.7.0rc1
Deprecated==1.2.10
distlib==0.3.1
distro==1.5.0
Django==3.1.7
django-jalali==4.1.0
html5lib==1.1
idna==2.5
ipaddr==2.2.0
isort==5.6.4
jdatetime==3.6.2
lazy-object-proxy==1.4.0
lockfile==0.12.2
lxml==4.6.2
Markdown==3.3.3
mccabe==0.6.1
mistune==0.8.1
msgpack==0.6.2
oauthlib==3.1.0
packaging==20.8
pep517==0.9.1
Pillow==8.1.2
progress==1.5
pycodestyle==2.6.0
pycparser==2.20
PyGithub==1.54.0
Pygments==2.7.3
PyJWT==1.7.1
pylint==2.6.0
PyNaCl==1.4.0
pyparsing==2.4.7
python3-openid==3.2.0
pytoml==0.1.21
pytz==2020.4
requests==2.20.0
requests-oauthlib==1.3.0
retrying==1.3.3
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
toml==0.10.2
urllib3==1.24.2
webencodings==0.5.1
wrapt==1.12.1
